Move valid_vectors.
