+++ cpp/tests/copying/nvstring_scatter_tests.cu
@@ -24,7 +24,7 @@
 #include <cudf/legacy/table.hpp>
 #include <random>
 #include <tests/utilities/nvcategory_utils.cuh>
-#include <tests/utilities/valid_vectors.h>
+#include <tests/utilities/legacy/valid_vectors.h>
 
 template <typename T>
 struct ScatterTest : GdfTest {};
