+++ cpp/tests/orderby/order_by_type_vectors.h
@@ -18,7 +18,7 @@
 #define ORDER_BY_TYPE_VECTORS_H
 
 #include <tests/utilities/legacy/tuple_vectors.h>
-#include <tests/utilities/valid_vectors.h>
+#include <tests/utilities/legacy/valid_vectors.h>
 
 #include <utilities/cudf_utils.h>
